Remove `TimeSigOverBkgPDFRatio`
In SkyLLH2 there is no pdf_type argument anymore.
The user needs to make sure that the correct PDFs are used.
So This class is not needed anymore in SkyLLH2.
We should remove it here.
Within the create_analysis function one could still specify the pdf_type argument if desired.