Add "physical" option to numerical stabilizer of energy bkg pdf
The numerical_stabilizer option for the energy bkg pdf for Public data
has been improved to allow to choose between a pure numerical solution
or a "physics" option based on astrophysical diffuse flux.

This variable is not anymore a boolean, but one of {None,'numerical','physical'}.

To allow control over the changing astrophysical diffuse flux parameters, the parameters
astrodiffplflux_phi0,astrodiffplflux_E0 and astrodiffplflux_gamma have been added to
the PDBakcgroundI3EnergyPDF, PDDatabackgroundI3EnergyPDF and the create_analysis instance
from time_integrated_stacked_ps.py.

The classes PDBakcgroundI3EnergyPDF and  PDDatabackgroundI3EnergyPDF now need a Dataset
instance to be initializiated if numerical_stabilizer='physics'.

The option only_one_pdf was additionally added to the class PDSignalEnergyPDFSetMultiSource
to allow a faster implementation of the physical stabilizer.