New astro diffuse bkg implementation for energy bkg pdf
The previous implementation included in commit de76f02 was wrong.
A new approach has been implemented, in which we create a function
that computes the expected signal from diffuse astrophysical bkg
and then add this to the energy bkg pdf computation for the
option numerical_stabilizer='physical'

Files with the precomputed results with latest data available at
this point for all datasets have been added. This avoids adding
too much time to create the analyses instance.

It is also possible to create files with personalized astrophysical
diffuse flux parameters to use in the analysis.