Poisson distribution at source level injection.
Correction for the event distribution in the injection for
several sources.

Previous method divived the n_injected events per source always the same
way for a given n_inj_total. Effectively this assume the most
favourable injecion in the trial.

The new method just uses a Poisson distribution for each source with
the expected mean. This gives more realistic injections.